Refactor technology registration and JSON encoding; streamline imports and enhance CustomJSONEncoder for heat generator classes
